+++ Smartscope/lib/montage.py
@@ -240,6 +240,8 @@ class BaseImage(ABC):
 
     def make_symlink(self):
         relative = os.path.relpath(self.raw,self.directory)
+        if self.image_path.exists():
+            return
         logger.debug(f'Relative path from {self.directory} to raw = {relative}')
         os.symlink(relative, self.image_path)
 
@@ -256,8 +258,8 @@ class Montage(BaseImage):
         super().__post_init__()
         self.directory.mkdir(exist_ok=True)
 
-    def load_or_process(self, check_AWS=False):
-        if self.check_metadata(check_AWS=check_AWS):
+    def load_or_process(self, check_AWS=False, force_process=False):
+        if not force_process and self.check_metadata(check_AWS=check_AWS):
             return
         self.metadata = parse_mdoc(self.mdoc, self.is_movie)
         self.build_montage()
@@ -364,15 +366,16 @@ class Target:
     stage_y: Union[float, None] = None
     stage_z: Union[float, None] = None
 
-    def __init__(self,shape: Union[list, Tensor], quality: Union[str,None]=None, from_center=False) -> None:
+    def __init__(self,shape: Union[list, np.array], quality: Union[str,None]=None, from_center=False) -> None:
         self.quality = quality
         if from_center:
             self.x = shape[0]
             self.y = shape[1]
             return
-        self.x = shape
-        self.y = shape
         self.shape = shape
+        self.x = None
+        self.y = None
+        
 
 
     @property
@@ -383,6 +386,10 @@ class Target:
     def x(self, value = None):
         if isinstance(value,list):
             self._x = int(value[0] + (value[2] - value[0]) // 2)
+            return 
+        if value is None:
+            self._x = int(self.shape[0] + (self.shape[2] - self.shape[0]) // 2)
+            return
         self._x = value    
 
     @property
@@ -393,10 +400,18 @@ class Target:
     def coords(self):
         return np.array([self._x,self._y])
 
+    @property
+    def stage_coords(self):
+        return np.array([self.stage_x,self.stage_y])
+
     @y.setter
     def y(self, value = None):
         if isinstance(value,list):
             self._y = int(value[1] + (value[3] - value[1]) // 2)
+            return
+        if value is None:
+            self._y = int(self.shape[1] + (self.shape[3] - self.shape[1]) // 2) 
+            return
         self._y = value
 
     def set_area_radius(self, shape_type):
@@ -404,17 +419,16 @@ class Target:
         len1 = int(self.shape[2] - self.shape[0])
         len2 = int(self.shape[3] - self.shape[1])
 
-        if shape_type == 'square':
-            self.area = len1 * len2
-            return
+        # if shape_type == 'square':
+        self.area = len1 * len2
+            # return
 
-        if shape_type == 'hole':
+        # if shape_type == 'hole':
 
-            self.radius = min(len1, len2) / 2
-            self.area = np.pi * (self.radius ** 2)
-            return
+        self.radius = min(len1, len2) / 2
+            # self.area = np.pi * (self.radius ** 2)
 
     def convert_image_coords_to_stage(self, montage):
-        tile, dist = closest_node([self.x, self.y], montage.metadata.piece_center)
+        tile, dist = closest_node(self.coords.reshape(-1,2), montage.metadata.piece_center)
         self.stage_x, self.stage_y = pixel_to_stage(dist, montage.metadata.iloc[tile], montage.metadata.iloc[tile].TiltAngle)
         self.stage_z = montage.stage_z
